ORY-1  #30min verilen sayıya kadar olan ciftleri bulan program
